+++ bluetooth_low_energy_platform_interface/lib/src/my_gatt_attribute.dart
@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 import 'gatt_attribute.dart';
 import 'uuid.dart';
 
@@ -8,4 +10,11 @@ abstract class MyGattAttribute implements GattAttribute {
   MyGattAttribute({
     required this.uuid,
   });
+}
+
+extension Uint8ListX on Uint8List {
+  Uint8List trimGATT() {
+    final elements = take(512).toList();
+    return Uint8List.fromList(elements);
+  }
 }