+++ bluetooth_low_energy_platform_interface/lib/src/my_gatt_characteristic.dart
@@ -9,14 +9,19 @@ class MyGattCharacteristic extends MyGattAttribute
     implements GattCharacteristic {
   @override
   final List<GattCharacteristicProperty> properties;
-  final Uint8List? value;
+  Uint8List? _value;
   @override
   final List<MyGattDescriptor> descriptors;
 
   MyGattCharacteristic({
     required super.uuid,
     required this.properties,
-    this.value,
+    Uint8List? value,
     required this.descriptors,
-  });
+  }) : _value = value?.trimGATT();
+
+  Uint8List get value => _value ?? Uint8List.fromList([]);
+  set value(Uint8List value) {
+    _value = value.trimGATT();
+  }
 }