+++ bluetooth_low_energy_platform_interface/lib/src/my_gatt_descriptor.dart
@@ -4,10 +4,15 @@ import 'gatt_descriptor.dart';
 import 'my_gatt_attribute.dart';
 
 class MyGattDescriptor extends MyGattAttribute implements GattDescriptor {
-  final Uint8List? value;
+  Uint8List? _value;
 
   MyGattDescriptor({
     required super.uuid,
-    this.value,
-  });
+    Uint8List? value,
+  }) : _value = value?.trimGATT();
+
+  Uint8List get value => _value ?? Uint8List.fromList([]);
+  set value(Uint8List value) {
+    _value = value.trimGATT();
+  }
 }